\name{01.Introduction.to.heR.Misc}

\alias{01.Introduction.to.heR.Misc}

\title{An Introduction to the heR.Misc Package}

\description{This help page gives a short introduction to the \code{heR.Misc}
package, which is a module of the Human Exposure Research (heR) package
}

\details{This package was first developed as part of Neil Klepeis' research
on human exposure to airborne
particulate matter occurring in people's homes while a PhD student in
Environmental Health Sciences at the School of Public Heatlh, University of
California at Berkeley.

The elements of this package include
basic functions for plotting and data analysis of human exposure-related data.

Please see the package index for a complete listing of the specific
functions and data sets that are part of this package.

The following heR modules currently exist:

\tabular{ll}{
heR.Activities \tab Human activity pattern plotting and analysis \cr
heR.ActivityData \tab Human activity pattern data \cr
heR.IndoorAir \tab Indoor air modeling \cr
heR.Inhalation \tab Inhalation exposure and dose modeling \cr
heR.Misc \tab Miscellaneous plotting and analysis functions \cr
heR.MonitoringData \tab Indoor and outdoor pollutant time series data \cr
heR.Simulation \tab Exposure simulation framework \cr
heR.SurveyData \tab PTEAM exposure survey data \cr
}

See \url{http://exposurescience.org/her.html} to access the software
and for further information on heR.

}

\section{Distribution License}{This package is distributed under the GNU General Public
License (GPL).  There is ABSOLUTELY NO WARRANTY for this software.

For more details visit
\url{http://www.gnu.org/copyleft/gpl.html} or write to:

The Free Software Foundation, Inc.,
59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.

In short, you may copy, modify, and redistribute the source code for
this package, so long as you make the complete source code available
for any version that you release -- and it is also licensed under the GPL.
In this way, the software remains \emph{free} for current and future
users.
}

\author{Neil E. Klepeis
\cr
\cr
\cr
\url{http://klepeis.net}
\url{http://exposurescience.org}
}

\keyword{documentation}


