\name{comments}

\alias{comments}

\title{Return Comments of Data Frame Columns (Variables)}


\description{A function to return the variable comments/labels for
a dataframe)
}


\usage{comments(z)
}


\arguments{
\item{z}{a dataframe object with embedded comments for each column (variable)}
}

\value{
A data frame with a column for variable names and a column for variable comments, i.e., labels.
}


\author{Neil E. Klepeis
\cr
\cr
\cr
\url{http://klepeis.net}
}


\keyword{datasets}



