\name{mond}

\alias{mond}

\title{Mondrian Plot}

\description{
Create a \emph{Mondrian} "World" plot for use in visualizing relationships
in an n-dimensional data space.  Useful for large data sets with many
records and multiple variates.
}

\usage{
mond(z, colors, levels=4, pretty=FALSE, dep=TRUE,
     nlab=4, samples=TRUE, plot=TRUE,
     legend=TRUE, xlab, ylab, main, ... )
}

\arguments{
  \item{z}{a dataframe object containing at least two variates (columns) and
         any number of records (rows)}
  \item{colors}{a vector of integers or color names, defaults to ("white", "black",
                "green", "red", "blue", ... ) }
  \item{levels}{number of splits (class intervals).  Can be an integer, a vector, or a list.
                A single integer indicates
                that all variates will be cut into that number of levels.  A vector (of at
                least length 2) specifies the number of cuts for each variate.  A list
                (with at least 2 components) specifies the exact N cut points (giving N-1
                different levels) for each variate. }
  \item{pretty}{a logical specifying whether to use levels generated by the
                either the pretty function (\code{pretty=TRUE}) or in even splits.  This argument is only relevant
                when \code{levels} is a single integer or a vector of integers.}
\item{dep}{logical, whether to plot variates as dependent or independent.  Currently unimplemented. }
  \item{nlab}{the number of tic marks and labels for the vertical axis}
  \item{samples}{a logical specifying whether or not the vertical axis
                 scales according to the number of samples (samples=TRUE) or the proportion of samples. }
  \item{legend}{A logical specifying whether or not a legend will be drawn
                for each variate (\code{legend=TRUE}).}
  \item{plot}{A logical specifying whether or not the plot will be drawn.  If
              \code{plot=FALSE}, then the function returns a list (see below).}
\item{xlab,ylab,main}{Axis labels and plot title}
\item{...}{additional graphical parameters}
}

\details{
A Mondrian or "World" plot is a simple sorted scatterplot of all the data
in a given dataframe.  It provides a quick "bird's eye" or "world" view of the
data space.  Colors are used to represent class intervals and the records
of sequential variates are sorted in a nested fashion to reveal conditional
dependencies.

The function makes use of existing R functions:
\cr
* \code{\link{cut}} -- to create coded levels in the dataframe \cr
* \code{\link{order}} -- to create a nested sort of the data space \cr
* \code{\link{image}} -- to plot the data samples (records) as a series of colored blocks;  Each block color (for each variate) corresponds to a specific class interval \cr
* \code{\link{hist}} -- to check that given cut points span the entire data range and to provide frequency counts to aid in the plotting of clarifying lines. \cr

Light gray lines are plotted that indicate
the number of samples (or the proportion of samples) that lie in a given
range of the root variate (i.e., a given class interval).

}

\value{
  When \code{plot=FALSE}, the function returns a list with the following components:
\cr
  \item{zord}{The sorted and recoded dataframe z.  This is the dataframe that is
plotted with the \code{image} function to produce the basic Mondrian plot.}
  \item{brks}{The breakpoints used for the root variate.}
  \item{cnts}{The counts in each class interval for the root variate.}
  \item{cumcnts}{The cumulative counts in each class interval for the root
variate.}
  \item{cumperc}{The cumulative proportion of counts in each class interval
for the root variate.}
}

\author{Neil E. Klepeis
\cr
\cr
\cr
\url{http://klepeis.net} \cr
\url{http://exposurescience.org}
}


\keyword{hplot}

