\name{peaks}

\alias{peaks}

\title{Locate peaks in a time series}

\description{A function to locate peaks in a time series
}


\usage{
peaks(series,span=3)
}


\arguments{
\item{series}{the time series}
\item{span}{?}
}


\details{
This is something Brian Ripley wrote that I thought was kinda
useful.
}


\value{
A vector showing the peaks
}


\author{Neil E. Klepeis
\cr 
\cr 
\cr
\url{http://klepeis.net}
}


\keyword{ts}



