\name{print.comments}

\alias{print.comments}

\title{Print Data Frame Comments}


\description{A function to print the variable comments/labels for
a dataframe and the overall object comment)
}


\usage{print.comments(z, add.lf=TRUE, show.levels=FALSE)
}


\arguments{
\item{z}{a dataframe object with embedded comments}
\item{add.lf}{logical, whether to add a line feed between each element of a comment vector}
\item{show.levels}{logical, whether to print out variable levels in addition to comments}
}



\author{Neil E. Klepeis
\cr
\cr
\cr
\url{http://klepeis.net}
}


\keyword{print}



