\name{sweights}
\alias{sweights}
\title{Calculate Stratification Weights}
\description{
Calculate per-record stratification or post-stratification 
probabilities for an arbitrary data set with pre-existing
per-record probabilities, and either equal or user-specified
group probabilities.


}
\usage{
sweights(f, w, tab, max.levels=25)
}
\arguments{
  \item{f}{a list or data frame containing equal length factors defining subgroups and corresponding to the elements of an arbitrary vector or the rows of an arbitrary data frame.}
  \item{w}{pre-existing weights for each record in \code{f}; records are assigned equal weights by default}
  \item{tab}{an optional dataframe containing weights for each unique combination of factors (subgroups), consisting of leading columns with names exactly corresponding to the factors in \code{f}, plus an additional column containing the actual weights. By default, equal weights are assigned to each subgroup.  This dataframe can be easily formed from a frequency array generated by \code{table} or \code{xtab}, e.g., by using the \code{as.data.frame.table} function.}
  \item{max.levels}{the maximum number of levels that are allowed for each factor, defaulting to 25; too many factors levels may signal that something is awry with the factor specification, e.g., not actually specifying the correct factor or specifying an improperly cut numeric vector}
} 
\details{
We calculate a weight for every record in the set of equal-length
factors in \code{f}.
The factors can correspond to any arbitrary data
object of the same length as the factors.

These per-record stratification weights for the data
set are calculated so that the probability of selecting a record in
a given subgroup (defined by factors in \code{f}, each the same
length as the unspecified data set) is equal to
desired probability as specified by
the flattened frequency table \code{tab}, defaulting to equal
probabilities for each
subgroup, if \code{tab} is not given.

The pre-existing probabilities for records, given by \code{w},
are preserved within the defined subgroups.
Pre-existing probabilities are
equalized across records, if \code{w} is not specified.

See more details on the calculation of weights below.

Note that the names of the dimensions in \code{tab} must match the names of the
factors in \code{f}.
Factor combinations in \code{f}, which don't have matches in \code{tab},
are assigned weights of zero.  
}

\section{Derivation of Per-Record Probability Factors}{
The pre-existing probabilities for each group, \eqn{i}{i}, are represented
by \eqn{P_i}{Pi} and calculated from the pre-exisiting per-record weights
\eqn{w}{w} as follows:

\deqn{P_i = sum{j}{n_i}{w_{ij}}}{Pi = sum(wij) from j = 1 to ni }

where \eqn{n_i}{ni} is the number of records in each group and \eqn{j}{j}
is the record index within a given group.

To obtain the desired per-group weights, represented as \eqn{p_i}{pi}, 
the pre-existing weights for records belonging to group \eqn{i}{i}
are multiplied by the ratio \eqn{\frac{p_i}{P_i}}{pi/Pi}.  After multiplication,
the sum of the new weights within each group will be \eqn{p_i}{pi}, and the
original relative weights for members of a group are unchanged.

For a simple example, let's assume that the original, pre-existing weights are
equal for all records and that the new group probabilities are to be equal
across all groups.  The total number of groups is \eqn{N}{N} and the total
number of records is \eqn{M}{M}, so that per-record probabilities are
\eqn{\frac{1}{M}}{1/M} and the group probabilites, \eqn{p_i}{pi},
are \eqn{\frac{1}{N}}{1/N}.
The pre-exisiting group
probabilities, \eqn{P_i}{Pi}, are equal to the number of records in a group
divided by the total number of records, \eqn{\frac{n_i}{M}}{ni/M}.  The factor to multiply by each member of a particular group \eqn{i}{i} is then:

\deqn{\frac{p_i}{P_i} = \frac{M}{N n_i}}{pi/Pi = M/(N*ni)}

Multiplying these factors by the per-record probability, \eqn{\frac{1}{M}}{1/M},
which is the same for all records in this example, 
the new per-record probabilities for each group are \eqn{\frac{1}{N n_i}}{1/(N*ni)}, which is simply the reciprocal of the product of the number of groups and the number of group records.
}
\value{
A vector of stratification (or post-stratification) weights equal in length to the factors in \code{f}.
}
\author{Neil Klepeis}
\keyword{misc}
