\name{tkdensity}

\alias{tkdensity}

\title{Tk Kernel Density}

\description{GUI function to interactively visualize
a kernel density estimate
}


\usage{
tkdensity()
}


\arguments{
}


\details{
Written as a \code{demo} by Thomas Lumley
}


\value{
None
}


\author{Neil E. Klepeis
\cr 
\cr 
\cr
\url{http://klepeis.net}
}


\keyword{nonparametric}




